+++ src/services/sign-in-audience.ts
@@ -1,5 +1,5 @@
-import { window, ThemeIcon, env, Uri, ThemeColor } from "vscode";
-import { SIGNIN_AUDIENCE_DOCUMENTATION_URI, SIGNIN_AUDIENCE_OPTIONS } from "../constants";
+import { window, ThemeIcon, ThemeColor } from "vscode";
+import { SIGNIN_AUDIENCE_OPTIONS } from "../constants";
 import { AppRegTreeDataProvider } from "../data/app-reg-tree-data-provider";
 import { AppRegItem } from "../models/app-reg-item";
 import { ServiceBase } from "./service-base";
@@ -8,8 +8,8 @@ import { GraphClient } from "../clients/graph-client";
 export class SignInAudienceService extends ServiceBase {
 
     // The constructor for the SignInAudienceService class.
-    constructor(treeDataProvider: AppRegTreeDataProvider, graphClient: GraphClient) {
-        super(treeDataProvider, graphClient);
+    constructor(graphClient: GraphClient, treeDataProvider: AppRegTreeDataProvider) {
+        super(graphClient, treeDataProvider);
     }
 
     // Edits the application sign in audience.
@@ -36,24 +36,8 @@ export class SignInAudienceService extends ServiceBase {
                 .then(() => {
                     this.triggerOnComplete({ success: true, statusBarHandle: status });
                 })
-                .catch(() => {
-                    if (item.contextValue! === "AUDIENCE-PARENT") {
-                        item.children![0].iconPath = new ThemeIcon("symbol-field", new ThemeColor("editor.foreground"));
-                    } else {
-                        item.iconPath = new ThemeIcon("symbol-field", new ThemeColor("editor.foreground"));
-                    }
-                    this.treeDataProvider.triggerOnDidChangeTreeData(item);
-                    status.dispose();
-
-                    window.showErrorMessage(
-                        `An error occurred while attempting to change the Sign In Audience. This is likely because some properties of the application are not supported by the new sign in audience. Please consult the Azure AD documentation for more information at ${SIGNIN_AUDIENCE_DOCUMENTATION_URI}.`,
-                        ...["OK", "Open Documentation"]
-                    )
-                        .then((answer) => {
-                            if (answer === "Open Documentation") {
-                                env.openExternal(Uri.parse(SIGNIN_AUDIENCE_DOCUMENTATION_URI));
-                            }
-                        });
+                .catch((error) => {
+                    this.triggerOnError({ success: false, statusBarHandle: status, error: error, treeViewItem: item.contextValue! === "AUDIENCE-PARENT" ? item.children![0] : item, previousIcon: new ThemeIcon("symbol-field", new ThemeColor("editor.foreground")) });
                 });
         }
     }